CSI/requirements.txt
fastapi==0.97.0
pandas~=2.0.1
scikit-learn~=1.2.2
datasets~=2.11.0
transformers==4.33.3
numpy==1.26.0
evaluate~=0.4.0
torch~=2.0.0
ray[tune]~=2.5.0
accelerate
protobuf==3.15.3
gradio~=3.34.0
uvicorn
pydantic~=1.10.9
PyYAML~=6.0